restructure and enhance DatePicker validation tests
