+++ vaadin-combo-box-flow-parent/vaadin-combo-box-flow/src/test/java/com/vaadin/flow/component/combobox/validation/MultiSelectComboBoxBasicValidationTest.java
@@ -50,12 +50,12 @@ public class MultiSelectComboBoxBasicValidationTest {
     @Test
     public void setInvalid_isInvalid() {
         Assert.assertFalse(testField.isInvalid());
-        Assert.assertFalse(testField.getElement().hasProperty("invalid"));
+        Assert.assertFalse(testField.getElement().getProperty("invalid", false));
 
         testField.setInvalid(true);
 
         Assert.assertTrue(testField.isInvalid());
-        Assert.assertTrue(testField.getElement().hasProperty("invalid"));
+        Assert.assertTrue(testField.getElement().getProperty("invalid", false));
     }
 }
