+++ vaadin-date-picker-flow-parent/vaadin-date-picker-flow/src/test/java/com/vaadin/flow/component/datepicker/validation/BasicValidationTest.java
@@ -13,21 +13,21 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.checkbox.tests.validation;
+package com.vaadin.flow.component.datepicker.validation;
 
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
 import com.vaadin.flow.component.HasValidation;
-import com.vaadin.flow.component.checkbox.CheckboxGroup;
+import com.vaadin.flow.component.datepicker.DatePicker;
 
-public class CheckboxGroupBasicValidationTest {
-    private CheckboxGroup<String> testField;
+public class BasicValidationTest {
+    private DatePicker testField;
 
     @Before
     public void setup() {
-        testField = new CheckboxGroup<>();
+        testField = new DatePicker();
     }
 
     @Test
