+++ vaadin-date-picker-flow-parent/vaadin-date-picker-flow/src/test/java/com/vaadin/flow/component/datepicker/validation/BinderValidationTest.java
@@ -1,5 +1,6 @@
-package com.vaadin.flow.component.datepicker;
+package com.vaadin.flow.component.datepicker.validation;
 
+import com.vaadin.flow.component.datepicker.DatePicker;
 import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.data.binder.BindingValidationStatus;
 import com.vaadin.flow.data.binder.BindingValidationStatusHandler;
@@ -14,7 +15,7 @@ import org.mockito.MockitoAnnotations;
 
 import java.time.LocalDate;
 
-public class DatePickerBinderValidationTest {
+public class BinderValidationTest {
 
     private static final String BINDER_FAIL_MESSAGE = "BINDER_FAIL_MESSAGE";
     private static final String BINDER_REQUIRED_MESSAGE = "REQUIRED";
