+++ vaadin-text-field-flow-parent/vaadin-text-field-flow/src/test/java/com/vaadin/flow/component/textfield/validation/AbstractBasicValidationTest.java
@@ -45,12 +45,12 @@ public abstract class AbstractBasicValidationTest<T extends Component & HasValid
     @Test
     public void setInvalid_isInvalid() {
         Assert.assertFalse(testField.isInvalid());
-        Assert.assertFalse(testField.getElement().hasProperty("invalid"));
+        Assert.assertFalse(testField.getElement().getProperty("invalid", false));
 
         testField.setInvalid(true);
 
         Assert.assertTrue(testField.isInvalid());
-        Assert.assertTrue(testField.getElement().hasProperty("invalid"));
+        Assert.assertTrue(testField.getElement().getProperty("invalid", false));
     }
 
     abstract protected T createTestField();
